separated the light classes for more shadow implementation, and fixing the lightcommonfactor address problem
