+++ CGImplementation/Graphics/CGAssetUtil.h
@@ -19,8 +19,8 @@ namespace CGProj
 		SHADER_DIR_VISUALIZER,
 		SHADER_POINT_VISUALIZER,
 		SHADER_SPOT_VISUALIZER,
-		SHADER_SHADOW_MAP,
-		SHADER_SHADOW_MAP_DEBUG_RENDER,
+		SHADER_DIR_SHADOW_MAP,
+		SHADER_DIR_SHADOW_MAP_DEBUG_RENDER,
 		NUM_CG_SHADER_ENUM
 	};
