+++ CGImplementation/Graphics/CGEditPrimitive.cpp
@@ -1,0 +1,207 @@
+#include <Graphics/CGEditPrimitive.h>
+
+// =================================================================
+/*** CG EDIT BOX    ***/
+CGProj::CGEditBox::CGEditBox()
+	:
+	m_BoxType(EDIT_PRIMITIVE_AABB),
+	m_position(glm::vec3(0)),
+	m_halfExtents(glm::vec3(1)),
+	m_orientation(glm::quat(0, 0, 0, -1))
+{
+	updateAABB();
+}
+
+CGProj::CGEditBox::CGEditBox(
+	const glm::vec3& pos, const glm::vec3& halfExtents, const glm::quat& m_orient)
+	: m_position(pos), m_halfExtents(halfExtents), m_orientation(m_orient)
+{
+	// halfExtents should be positive
+	assert(halfExtents.x > 0 || halfExtents.y > 0 || halfExtents.z > 0);
+	m_BoxType = EditPrimitiveType::EDIT_PRIMITIVE_AABB;
+	updateAABB();
+}
+
+void CGProj::CGEditBox::setPosition(const glm::vec3 & p)
+{
+	CGEditBox::m_position = p;
+	updateAABB();
+}
+
+void CGProj::CGEditBox::setPosition(const GPED::real x, const GPED::real y, const GPED::real z)
+{
+	m_position.x = x;
+	m_position.y = y;
+	m_position.z = z;
+	updateAABB();
+}
+
+void CGProj::CGEditBox::setXposition(const GPED::real x)
+{
+	m_position.x = x;
+	updateAABB();
+}
+
+void CGProj::CGEditBox::setYposition(const GPED::real y)
+{
+	m_position.y = y;
+	updateAABB();
+}
+
+void CGProj::CGEditBox::setZposition(const GPED::real z)
+{
+	m_position.z = z;
+	updateAABB();
+}
+
+glm::vec3 CGProj::CGEditBox::getPosition()
+{
+	return m_position;
+}
+
+void CGProj::CGEditBox::setHalfSize(const glm::vec3 & h)
+{
+	m_halfExtents = h;
+	updateAABB();
+}
+
+void CGProj::CGEditBox::setHalfSize(const GPED::real x, GPED::real y, GPED::real z)
+{
+	m_halfExtents = glm::vec3(x, y, z);
+	updateAABB();
+}
+
+void CGProj::CGEditBox::setXHalfSize(const GPED::real x)
+{
+	m_halfExtents.x = x;
+	updateAABB();
+}
+
+void CGProj::CGEditBox::setYHalfSize(const GPED::real y)
+{
+	m_halfExtents.y = y;
+	updateAABB();
+}
+
+void CGProj::CGEditBox::setZHalfSize(const GPED::real z)
+{
+	m_halfExtents.z = z;
+	updateAABB();
+}
+
+glm::vec3 CGProj::CGEditBox::getHalfSize()
+{
+	return m_halfExtents;
+}
+
+GPED::c3AABB CGProj::CGEditBox::getFitAABB()
+{
+	return m_fitAABB;
+}
+
+void CGProj::CGEditBox::setPrimitiveType(EditPrimitiveType e)
+{
+	assert(e == EDIT_PRIMITIVE_AABB || e == EDIT_PRIMITIVE_OBB);
+	m_BoxType = e;
+	updateAABB();
+}
+
+CGProj::EditPrimitiveType CGProj::CGEditBox::getPrimitiveType()
+{
+	return m_BoxType;
+}
+
+void CGProj::CGEditBox::updateAABB()
+{
+	switch (m_BoxType)
+	{
+	case EDIT_PRIMITIVE_AABB:
+		m_fitAABB = GPED::makeAABB(m_position, m_halfExtents);
+		break;
+	case EDIT_PRIMITIVE_OBB:
+		m_fitAABB = GPED::makeAABB(glm::mat3_cast(m_orientation), m_position, m_halfExtents);
+		break;
+	default:
+		assert(0); // fail to set the correct primitive type for box
+		break;
+	}
+}
+/*** CG EDIT BOX  ***/
+// =================================================================
+
+// =================================================================
+/*** CG EDIT Sphere  ***/
+CGProj::CGEditSpere::CGEditSpere()
+	:
+	m_position(glm::vec3(0)),
+	m_radius(1)
+{
+	updateAABB();
+}
+
+CGProj::CGEditSpere::CGEditSpere(const glm::vec3 & pos, const GPED::real radius)
+	: m_position(pos), m_radius(radius)
+{
+	assert(radius > 0);
+	updateAABB();
+}
+
+void CGProj::CGEditSpere::setPosition(const glm::vec3 & p)
+{
+	CGEditSpere::m_position = p;
+	updateAABB();
+}
+
+void CGProj::CGEditSpere::setPosition(const GPED::real x, const GPED::real y, const GPED::real z)
+{
+	m_position.x = x;
+	m_position.y = y;
+	m_position.z = z;
+	updateAABB();
+}
+
+void CGProj::CGEditSpere::setXposition(const GPED::real x)
+{
+	m_position.x = x;
+	updateAABB();
+}
+
+void CGProj::CGEditSpere::setYposition(const GPED::real y)
+{
+	m_position.y = y;
+	updateAABB();
+}
+
+void CGProj::CGEditSpere::setZposition(const GPED::real z)
+{
+	m_position.z = z;
+	updateAABB();
+}
+
+glm::vec3 CGProj::CGEditSpere::getPosition()
+{
+	return m_position;
+}
+
+void CGProj::CGEditSpere::setRaidus(GPED::real r)
+{
+	CGEditSpere::m_radius = r;
+	updateAABB();
+}
+
+GPED::real CGProj::CGEditSpere::getRadius()
+{
+	return m_radius;
+}
+
+GPED::c3AABB CGProj::CGEditSpere::getFitAABB()
+{
+	return m_fitAABB;
+}
+
+void CGProj::CGEditSpere::updateAABB()
+{
+	m_fitAABB = GPED::makeAABB(m_position, m_radius);
+}
+/*** CG EDIT Sphere  ***/
+// =================================================================
