+++ CGImplementation/ShaderFolder/CGDirDepthMap.vs
