+++ CGImplementation/ShaderFolder/CGDirDepthMapDebugRender.vs
